+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### Unreleased
 
+* Add `Railsboot::ContainerComponent`
+
 ### 0.1.0
 
 * Initial release